requirements.txt
langchain-google-genai
streamlit
duckduckgo-search
python-dotenv
